requirements.txt
asgiref==3.4.1
backcall==0.2.0
black==21.9b0
click==8.0.1
decorator==5.1.0
Django==3.2.7
djangorestframework==3.12.4
flake8==3.9.2
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pytz==2021.3
regex==2021.9.30
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
typing-extensions==3.10.0.2
wcwidth==0.2.5
